Reorganized app into a package.
Updating image paths in __init__.py and in test_app.py for create and edit routes because now static is in a different location relative to where these files are ran. Updated all imports to accomodate for the new package setup.
