+++ application/static/post_imgs/3/Denver's Racial Distribution.png
