+++ application/templates/admin.html
