+++ bugs.txt
@@ -27,6 +27,7 @@ Bugs:
 - migrate db to aware datetimes with an hour
 - get rid of the hardcoded env for the app caches of css
 - remove flask env from default config - have it set as env var
+- old images are not removed from file system if the post is edited to get rid of them...is this how I want things to work?
 
 Problems with the current app testing flow:
 - for subscribe/create tests, if they fail the item is still left in the database. this will be fixed with app factory setup. thank goodness.
