+++ tests/unittests/test_app_factory.py
@@ -1,4 +1,4 @@
-from app import create_app
+from application import create_app
 
 """Tests the app factory setup."""
